move shared fixtures to conftest to avoid code duplication
